Profil kezelésének véglegesítésénel feltöltése.

+++ posts/delete.php
@@ -1,33 +1,34 @@
 <?php
 Use App\Tools;
-require_once __DIR__ . '/../lib/autoload.php';
-$postItem = new \App\Models\Posts;
-$postItem->getItemBy('slug',$_GET['slug']);
+requireonce _DIR . '/../lib/autoload.php';
+$postItem = new App\Models\Posts;
+$comments = new App\Models\Comment;
+$recoms = new App\Models\Recommend;
+$decoms = new App\Models\Unrecommend;
+$db = new App\Database;
 
-$comments = new \App\Models\Comment;
-$commets = $comments->getItemBy('post_id',$postItem->id);
-
-$db= new App\Database;
-$liked_post=$db->getItemByValue('recommended_post','post_id',$postItem->id);
-if($liked_post)
-{
-    $db->delete('recommended_post',$liked_post[0]["id"]);
-
+$postItem = $postItem->getItemBy('slug',$_GET['slug']);
+$comments = $comments->getItemsBy('post_id', $postItem->id);
+$recoms = $recoms->getItemsBy('post_id', $postItem->id);
+$decoms = $decoms->getItemsBy('post_id', $postItem->id);
+try{
+    Image::RemoveUploadedImage('/files/blog_image/'.$postItem->image);
+    if($decoms){
+        $db->delete_all('unrecommended_post','post_id', $postItem->id);
+    }
+    if($recoms){
+        $db->delete_all('recommended_post','post_id', $postItem->id);
+    }
+    if($comments){
+        $db->delete_all('comment','post_id', $postItem->id);
+    }
+    if($postItem){
+        $postItem->delete();
+    }
+    Tools::FlashMessage('Sikeresen Törölve lett','success');
+    header('Location:/');
+}catch(Exception $e){
+    Tools::FlashMessage('Hiba: '.$e,'danger');
+    header('Location: /');
 }
-$disliked_post=$db->getItemByValue('unrecommended_post','post_id',$postItem->id);
-if($disliked_post)
-{
-    $db->delete('unrecommended_post',$disliked_post[0]["id"]);
-
-}
-if($comments){
-    $comments->delete_all($postItem->id);
-
-}
-
-$postItem->delete();
-
-Tools::FlashMessage('Sikeresen Törölve lett','success');
-header('Location:/');
-
 ?>
